+++ src/main/java/com/PuzzleU/Server/heart/Service/HeartService.java
@@ -75,4 +75,5 @@ public class HeartService {
         competitionRepository.save(competition);
         return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "좋아요를 취소하였습니다"), null);
     }
+
 }
