+++ src/main/java/com/PuzzleU/Server/team/service/TeamService.java
@@ -47,6 +47,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -602,5 +603,200 @@ public class TeamService {
         else {
             throw new RestApiException(ErrorType.NOT_MATCH_ACCEPT_REJECT);
         }
+    }
+    @Transactional
+    public ApiResponseDto<TeamListDto> getTeamList(Long competition_Id, int pageNo, int pageSize, String sortBy) {
+        try {
+            TeamListDto teamListDto = new TeamListDto();
+            Pageable pageable = PageRequest.of(pageNo, pageSize, Sort.by(sortBy).descending());
+
+            // 경기 정보 찾기
+            Competition competition = competitionRepository.findById(competition_Id)
+                    .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COMPETITION));
+
+            // 해당 경기에 속한 팀들 가져오기
+            Page<Team> teamPage;
+            teamPage = new PageImpl<>(teamRepository.findByCompetition(competition, pageable));
+            // 팀 목록 및 팀 정보 DTO 작성
+            List<TeamAbstractBaseDto> teamAbstractDtos = teamPage.getContent().stream()
+                    .map(team -> {
+                        TeamAbstractBaseDto teamAbstractDto = new TeamAbstractBaseDto();
+                        // 팀 작성자 설정
+                        List<TeamUserRelation> teamUserRelation = teamUserRepository.findByTeam(team);
+                        for (TeamUserRelation teamuserRelation : teamUserRelation) {
+                            if (teamuserRelation.getIsWriter()) {
+                                teamAbstractDto.setTeamWriter(teamuserRelation.getUser().getUserKoreaName());
+                                break;
+                            }
+                        }
+                        // 팀 위치 설정
+                        List<TeamLocationRelation> teamLocationRelation = teamLocationRelationRepository.findByTeam(team);
+                        List<String> locationList = new ArrayList<>();
+                        for (TeamLocationRelation teamLocationRelation1 : teamLocationRelation) {
+                            String location = teamLocationRelation1.getLocation().getLocationName();
+                            locationList.add(location);
+                        }
+                        teamAbstractDto.setTeamId(team.getTeamId());
+                        teamAbstractDto.setTeamNeed(team.getTeamMemberNeed());
+                        teamAbstractDto.setTeamNowMember(team.getTeamMemberNow());
+                        teamAbstractDto.setTeamTitle(team.getTeamTitle());
+                        teamAbstractDto.setTeamPoster(competition.getCompetitionPoster());
+                        teamAbstractDto.setTeamLocations(locationList);
+                        List<Position> positionLists= teamPositionRelationRepository.findByTeam(team);
+                        // 포지션 설정
+                        List<String> positionList = positionLists.stream()
+                                .map(Position::getPositionName)
+                                .collect(Collectors.toList());
+                        teamAbstractDto.setPositionList(positionList);
+                        return teamAbstractDto;
+                    })
+                    .collect(Collectors.toList());
+
+            // 팀 목록 DTO 설정
+            teamListDto.setTotalTeam((int) teamPage.getTotalElements());
+            teamListDto.setTotalElements(teamPage.getTotalElements());
+            teamListDto.setTeamList(teamAbstractDtos);
+            teamListDto.setLast(teamPage.isLast());
+            teamListDto.setTotalPages(teamPage.getTotalPages());
+            teamListDto.setPageNo(pageNo);
+            teamListDto.setPageSize(pageSize);
+
+            return ResponseUtils.ok(teamListDto, null);
+        } catch (RestApiException e) {
+            // 이미 정의된 예외인 경우 그대로 던집니다.
+            throw e;
+        } catch (NoSuchElementException e) {
+            // 경기 또는 팀이 존재하지 않는 경우 예외 처리합니다.
+            throw new RestApiException(ErrorType.NOT_FOUND);
+        } catch (Exception e) {
+            // 서버 오류로 처리합니다.
+            throw new RestApiException(ErrorType.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @Transactional
+    public ApiResponseDto<TeamSpecificDto> getTeamSpecific(Long competition_Id, Long team_Id) {
+        try {
+            TeamSpecificDto teamSpecificDto = new TeamSpecificDto();
+            Competition competition = competitionRepository.findById(competition_Id)
+                    .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COMPETITION));
+
+            Team team = teamRepository.findById(team_Id)
+                    .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_TEAM));
+
+            // 팀 작성자 설정
+
+            List<TeamUserRelation> writerList = teamUserRepository.findByTeam(team);
+            if (!writerList.isEmpty()) {
+                for(TeamUserRelation teamUserRelation : writerList) {
+                    if(teamUserRelation.getIsWriter()) {
+                        teamSpecificDto.setTeamWriter(teamUserRelation.getUser().getUserKoreaName());
+                        break;
+                    }
+                }
+            } else {
+                throw new RestApiException(ErrorType.NOT_FOUND);
+            }
+
+
+
+            // 팀 위치 설정
+            List<TeamLocationRelation> teamLocationRelation = teamLocationRelationRepository.findByTeam(team);
+            List<String> locationList = new ArrayList<>();
+            for (TeamLocationRelation teamLocationRelation1 : teamLocationRelation) {
+                String location = teamLocationRelation1.getLocation().getLocationName();
+                locationList.add(location);
+            }
+            teamSpecificDto.setTeamNeed(team.getTeamMemberNeed());
+            teamSpecificDto.setTeamNowMember(team.getTeamMemberNow());
+            teamSpecificDto.setTeamTitle(team.getTeamTitle());
+            teamSpecificDto.setTeamPoster(competition.getCompetitionPoster());
+            teamSpecificDto.setTeamLocations(locationList);
+            List<Position> positionList= teamPositionRelationRepository.findByTeam(team);
+            // 포지션 설정
+            List<String> PositionList = positionList.stream()
+                    .map(Position::getPositionName)
+                    .collect(Collectors.toList());
+            teamSpecificDto.setPositionList(PositionList);
+
+            teamSpecificDto.setTeamContent(team.getTeamContent());
+            teamSpecificDto.setTeamIntroduce(team.getTeamIntroduce());
+
+            // 팀 멤버 설정
+            List<UserSimpleDto> userSimpleDtoList = new ArrayList<>();
+            List<TeamUserRelation> teamUserRelations = teamUserRepository.findByTeam(team);
+            for (TeamUserRelation teamUserRelation1 : teamUserRelations) {
+                UserSimpleDto userSimpleDto = new UserSimpleDto();
+                userSimpleDto.setUserId(teamUserRelation1.getUser().getId());
+                userSimpleDto.setUserProfile(teamUserRelation1.getUser().getUserProfile());
+                userSimpleDto.setUserKoreaName(teamUserRelation1.getUser().getUserKoreaName());
+                userSimpleDto.setUserRepresentativeProfileSentence(teamUserRelation1.getUser().getUserRepresentativeProfileSentence());
+                userSimpleDtoList.add(userSimpleDto);
+            }
+            teamSpecificDto.setMembers(userSimpleDtoList);
+
+            return ResponseUtils.ok(teamSpecificDto, null);
+        } catch (RestApiException e) {
+            throw e; // 이미 정의된 예외인 경우 그대로 던집니다.
+        } catch (Exception e) {
+            throw new RestApiException(ErrorType.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    // 팀을 검색했을때 정보 얻기
+    @Transactional
+    public ApiResponseDto<TeamListDto> getTeamSearchList(int pageNo, int pageSize, String sortBy, String search)
+
+    {
+        TeamListDto teamListDto = new TeamListDto();
+        Pageable pageable = PageRequest.of(pageNo, pageSize, Sort.by(sortBy).descending());
+        Page<Team> teamPage;
+
+        teamPage = new PageImpl<>(teamRepository.findByTeamTitleContaining(search, pageable));
+        List<TeamAbstractBaseDto> teamAbstractDtos;
+        teamAbstractDtos = teamPage.getContent().stream()
+                .map(team ->
+                {
+                    TeamAbstractBaseDto teamAbstractDto = new TeamAbstractBaseDto();
+                    List<TeamUserRelation> teamUserRelation = teamUserRepository.findByTeam(team);
+                    for (TeamUserRelation teamuserRelation : teamUserRelation) {
+                        if (teamuserRelation.getIsWriter())
+                        {
+                            System.out.println(teamuserRelation.getIsWriter());
+                            System.out.println(teamuserRelation.getUser().getUserKoreaName());
+                            teamAbstractDto.setTeamWriter(teamuserRelation.getUser().getUserKoreaName());
+                            break;
+                        }
+                    }
+                    List<TeamLocationRelation> teamLocationRelation = teamLocationRelationRepository.findByTeam(team);
+                    List<String> locationList = new ArrayList<>();
+                    for (TeamLocationRelation teamLocationRelation1 : teamLocationRelation) {
+                        String location = teamLocationRelation1.getLocation().getLocationName();
+                        locationList.add(location);
+                    }
+                    teamAbstractDto.setTeamId(team.getTeamId());
+                    teamAbstractDto.setTeamNeed(team.getTeamMemberNeed());
+                    teamAbstractDto.setTeamNowMember(team.getTeamMemberNow());
+                    teamAbstractDto.setTeamTitle(team.getTeamTitle());
+                    teamAbstractDto.setTeamPoster(team.getCompetition().getCompetitionPoster());
+                    teamAbstractDto.setTeamLocations(locationList);
+                    List<String> PositionList = new ArrayList<>();
+                    List<Position> positionList= teamPositionRelationRepository.findByTeam(team);
+                    for(Position position : positionList)
+                    {
+                        PositionList.add(position.getPositionName());
+                    }
+                    teamAbstractDto.setPositionList(PositionList);
+                    return teamAbstractDto;
+                })
+                .collect(Collectors.toList());
+        teamListDto.setTeamList(teamAbstractDtos);
+        teamListDto.setLast(teamPage.isLast());
+        teamListDto.setTotalPages(teamPage.getTotalPages());
+        teamListDto.setTotalElements(teamPage.getTotalElements());
+        teamListDto.setPageNo(pageNo);
+        teamListDto.setPageSize(pageSize);
+
+        return ResponseUtils.ok(teamListDto, null);
     }
 }
